py_backend/__main__.py
from __future__ import annotations

from pathlib import Path


def _maybe_load_dotenv() -> None:
    try:
        from dotenv import load_dotenv  # type: ignore
    except Exception:
        return

    repo_root = Path(__file__).resolve().parents[1]
    dotenv_path = repo_root / ".env"
    if dotenv_path.exists():
        load_dotenv(dotenv_path=dotenv_path, override=False)


_maybe_load_dotenv()

from .main import run


if __name__ == "__main__":
    run()
