requirements.txt
starlette==0.45.3
uvicorn==0.34.0
httpx==0.27.2
aiomysql==0.2.0
APScheduler==3.10.4
python-dotenv==1.0.1

# dev/test
pytest==8.3.4
pytest-asyncio==0.25.3
respx==0.22.0
